requirements.txt
requests
scrapy
selenium
python-dotenv
thefuzz
click
Flask
Flask-Cors